split cc factory into default and server-only cc factory
Summary: This diff creates a new `ServerCongestionControllerFactory` that will be used to create cc instances on the server only and thus can have different dependencies. At the moment this mirrors `DefaultCongestionControllerFactory`, but later in the stack when CCP is added, only the server cc factory will depend on (and be able to create) CCP instances, but the default one will not. This prevents any client builds from depending on CCP.

Reviewed By: yangchi

Differential Revision: D22139289

fbshipit-source-id: 2987f0234bc54ea6101ca0030b319b460571adeb