add handoff function to QuicCubic
Summary:
If mvfst hasn't heard from CCP in a while (e.g. if CCP crashed), we want to fall back to a default cc algorithm rather than stalling and waiting. However, if we create a new cc instance for all of the current connections, they will all act as if they are new connections and e.g. go through slow start again, which would (a) temporarily decrease the sending rate and (b) create a huge burst of all the connections going through slow start at once.

Instead, we can try our best to start the default algorithm where ccp left off, by passing along the last cwnd ccp reported (and potentially other information), and making sure the  algorithm does not go through slow start.

This diff adds the `handoff` function to implement this. For now I've only added it to cubic because it's the simplest case (BBR and Copa have a lot more state that needs to be figured out), but it could be added to these other algorithms later if desirable.

Once the QuicCCP cc algorithm is added (D21854339), it will be invoked as follows:
* the QuicCCP cc algorithm will internally maintain an instance of QuicCubic (or any other cc)
* libccp internally maintains a timer (one per serverworker) that keeps track of how long it has been since any connections have received a response from ccp.
* on each ack, when we call into libccp, it checks this timer. if the timer has surpassed a timeout value (for any connection, not just this one, because we may not have sent any messages yet), we change to "fallback mode" and call cubic:handoff. alternatively, if we are currently in fallback mode but libccp says the timer has been reset, we change back to "normal mode" and pick up where cubic left off by setting `cwnd = cubic::getCongestionWindow()`
* on future acks, if we are in fallback mode, we proxy the `onPacketAckOrLoss` calls to our internal cubic. however we still continue to update our own state so that we still have all of the connection updates to pass to ccp if it's eventually restored.

Reviewed By: yangchi

Differential Revision: D21854337

fbshipit-source-id: d0ef2b8da1d0af4faa2ae80b8c4f0ffb1c63be92