Add persistent congestion TODO.
Summary:
There's a lot of discussion about this in the WG. See:

https://github.com/quicwg/base-drafts/issues/3937

https://github.com/quicwg/base-drafts/issues/3939

https://github.com/quicwg/base-drafts/issues/3875

Our approach is very conservative, which may mean we are declaring persistent congestion mor frequently than we need to. This seems to work okay, but put a TODO to remind us to experiment with this in the future.

Reviewed By: yangchi

Differential Revision: D22699664

fbshipit-source-id: c0b73fc48a9f3dd141cada95eeb8493d17322702