Collect some tokens on refresh.
Summary:
This is a slightly more principled and perhaps less aggressive version of what we were doing before. The idea is as follows:

lastWriteTime represents the last time we consumed some amount of tokens. Assume that between writes we refresh the pacing rate. If we set tokens back to the batch size, then we will essentially eliminate any collecting that should have happened in the interim.

This diff compensates for this by first using lastWriteTime to collect tokens at the old rate, then setting lastWriteTime to now, so that the new rate will be used to collect any further tokens on the next write.

Reviewed By: yangchi

Differential Revision: D22669650

fbshipit-source-id: 2e937d9ede8c2f521613392fe44719fc82924a9b