Remove use of undefined return type
Summary:
When `CCP_ENABLED` isn't set, `CCPReader` stubs out the return type of
the `getDatapath()` with an undefined forward declaration of the missing
struct, which causes errors in modular builds, as it's never defined (and
so the modules backend can't figure out that the method declaration and
out-of-line definition's return types refer to the same type).

As this method is never used w/o `CCP_ENABLED`, just `#ifdef` out the helper
method altogether.

Reviewed By: igorsugak

Differential Revision: D22808308

fbshipit-source-id: 190427639343e8f7f4b6e1de78e7da7da1590968