Introduce size-enforced packet builder
Summary: As part of the d6d algorithm, we need to send probe packets of specified size and they are greater than the current pmtu. Introduce ~~SizeEnforcedPacketBuilder~~ RegularSizeEnforcedPacketBuilder for chained memory and InplaceSizeEnforcedPacketBuilder for continuous memory. They are ~~which is a~~ thin wrappers that append~~s~~ padding frames on top of the body of an existing packet. It will only force packet size if the packet has short header, since ~~I think it doesn't make sense~~ the spec does not allow ~~to start~~ probing pmtu until handshake completes.

Reviewed By: yangchi

Differential Revision: D22365558

fbshipit-source-id: 100118da9f2d1ed8b1234e36d1d271225dfa11b0