Add WindowedCounter
Summary: This implements a simple WindowedCounter, where given a `window` and `threshold` (which is meant to be small, e.g. < 10), it keeps track of a moving window and a count of samples with timestamps within that window. This will be useful for keeping track of consecutive loss of big packets, which in turn signals PMTU blackhole.

Reviewed By: mjoras

Differential Revision: D23534318

fbshipit-source-id: 59b21315f718a95846010995da16a4173b172bb2