Allow adjustment of RTT used for pacing calculation.
Summary:
This is effectively a way to increase/decrease the pacing rate. This can compesnate for the fact that pacing evenly over an entire RTT may be too conservative to utilize bandwidth fully. This is only used for the TokenlessPacer right now.

For BBR, pace over 50% of the RTT during startup, and 80% the rest of the time, as an arbitrary first attempt.

Reviewed By: yangchi

Differential Revision: D23733790

fbshipit-source-id: 54ccb49345eba9f50f8c1ca357984667bcea06a2