Calculate and send goodput estimate to client on httpTransactionDone
Summary:
When every HTTP transaction is complete, we test whether the connection is
suitable for sending a goodput knob frame. If we decide based on the following
conditions that the connection is suitable:
1. Number of "active" HTTP transactions currently on the connection is 0
2. Connection's transport type is QUIC, not TCP.
3. Connection has been upsampled for the GOODPUT_KNOB_FRAME sampling decision.

We will immediately compute the
connection level goodput using all the HTTP transactions that have been
captured up until this point in the connection and send the goodput estimate as
a QUIC knob frame to the client.

Reviewed By: bschlinker

Differential Revision: D25444649

fbshipit-source-id: 5867ce2fe289a899ea91d6431f7fde10f74e6af5