Create a unified Observer callback class, merge LifecycleObserver and InstrumentationObserver into it
Summary:
Here, we introduce a new Observer class that is a container for all the
callbacks that the QUIC Socket layer will invoke when interesting events happen
during the lifetime of the socket.

This class contains a union of all callbacks that are present in the
LifecycleObserver and InstrumentationObserver classes.

This is done in an effort to merge both the callback classes into one,
for easier maintainability. LifecycleObserver and InstrumentationObserver
are left behind temporarily, they will be deprecated in a future diff (up the stack).

Reviewed By: bschlinker

Differential Revision: D25845835

fbshipit-source-id: 425c865146ca061063cf6e463f8da92395f119f2