Use ackedBytes instead of packet size when no previous ACK sample.
Summary:
The previous logic didn't make a ton of sense. It would always take a bandwidth sample in the range [minPacketSize, maxPacketSize], when as the comment notes there could be multiple packets ACKed. Instead, just use the ackedBytes since that's at least somewhat principled.

This especially hurts 0RTT as when the 1RTT ack for the server's first app data flight comes in there will be no previous ACK from the client.

Reviewed By: yangchi

Differential Revision: D26375563

fbshipit-source-id: 37ffb1df3a2dfbd5a8b2d18ff19737b136defa86