Always set stateless reset token.
Summary:
This is another bandaid around our incorrect not updating of 0-RTT transport parameters on the client. The long term fix requires us to update all the parameters based on the final ones from the server. This will require a bit of a refactor to this area of the code as well.

Until we do that, manually set the stateless reset token so that 0-RTT clients can actually parse server stateless resets.

Reviewed By: yangchi

Differential Revision: D27875279

fbshipit-source-id: 1d08a53f2a9d876635dcd884b6bba316224543dd