Revert the fix for duplicate byte event registrations
Summary:
Diff D27846813 (https://github.com/facebookincubator/mvfst/commit/98b5e2aa8f2ab5bf035e73c1a99f86d46e740c8a) introduced a fix for a corner case in the ByteEventRegistration logic where-in it was possible to register a byte event for the same stream ID, offset and ByteEvent type as long as they were not consecutive registrations. If there was a different byte event registration for the same stream ID in-between, then duplicates were being allowed.
However, that fix causes an unexpected crash in HQSession::dropConnectionSync().

This commit reverts the fix for duplicate byte event registrations until we are able to find the root cause for the crash and come up with a better fix. Note that we are still able to catch duplicate registrations as long as they are consecutive registrations for the same <stream ID, offset, callback>. With this fix reverted, we will not be able to catch non-consecutive duplicate registrations.

Reviewed By: bschlinker

Differential Revision: D27903179

fbshipit-source-id: b4da3128d5ba14fc764e1768fd27982553968997