Allow writeChain to write pure EOM after BufferMeta is already written to a
Summary:
The CHECK is too harsh today. It blocks any write if BufferMeta has
been written to a Quic stream. Writing a pure EOM should be allowed. If caller
wants to write just an EOM, it should be free to choose either writeChain or
writeBufMeta API.

(Note: this ignores all push blocking failures!)

Reviewed By: mjoras

Differential Revision: D27684033

fbshipit-source-id: 475c65c7eac110ca6eb2a8c2abaf1eb00628a492