The DSR Sender is per-stream
Summary:
some of the code was written when we thought we'd have a
per-connection sender. Now that the sender is per-stream, these code are
updated. It is a bit awkward though. The write() function doesn't know what
stream it will be writing. Scheduler knows it, but Scheduler isn't the right
place to decide when to flush() the sender. So now the return type of packet
scheduling will include a raw pointer to the sender.

Reviewed By: mjoras

Differential Revision: D27869558

fbshipit-source-id: 2490d364d79d22db6f66eb82b194cfd51c182844