Circular deque
Summary:
This is a data structure where the storage is one
continuous piece, as opposed to std::deque where the
storage *may* be multiple fixed size pieces. There is
a similar data structure in Chromium. That's where I
borrowed the name.

We should just copy Chromium's circular deque into
Mvfst's third-party directory if it's just a single header.
Unfortunately it isn't.

Similar to std::deque it can push/pop from both ends.

In QUIC We have some use case that's currently using
std::deque, but can be better off with a circular deque.
Datagram is one such case.

This circular deque uses std::vector as underlying storage.
It isn't exactly API equivalent to std::deque since it doesn't
support custom Allocator, and i didn't implement a couple
APIs. But the rest are the same, or I tried to make them
same. We don't really use the missing APIs in Mvfst so
at least the transition will be straightforward.

Operations at both ends are 5-10% faster than deque.
But anything in the middle can be 2X slower. In practice
it won't be that bad if the "middle" is very close to either
end. I documented some number in the
CircularDequeBench.cpp file.

Internal to the CircularDeque, there are two indices `begin_`
and `end_` tracks where the data are in the underlying
vector. If they are at the same position, we treat the
CircularDeque as empty. Since it's circular, these indices
can wrap around the end of vector. For this reason, in a
not empty CircularDeque, they should never meet. Since
`end_` points to the one position beyond last legit element,
this also means it may occupy a legit position in the
underlying vector. So one surprising piece of code is
resize() secretly reserves S+1 in the vector when user
passes in S as the desired capacity.

The iterators also need to deal with the wrapping around
property. An iterator cannot stay at `storage_.end()`
position if the data has wrapped around the `end()`.
But it can stay there if the isn't wrapping around `end()`.

It's hard for me to tell if having all these branches hurt
the performance more, or if having multiple pieces of
storages like in std::deque hurt it more. Some very
simplified benchmark looks all right.

I thought about how to deal with the perf problem in
erase() from the middle. There are 3 options:

(1) Do not support it.
(2) Keep the hole there, do not move the rest of the
data to keep them continuous. Then use a
quic::IntervalSet to keep track what indices is a
"hole" and skip them in any type of access.
(3) Same as (2), except use a SkipList to track the holes
which can be slightly faster than IntervalSet.

But then neither (2) and (3) also solves the perf problem
of insert()/emplace() in the middle. Unless there happens
to be a "hole" next to the insert position.

Not supporting insert/erase in the middle is very
attractive. But then we will be further away from API
equivalent to a std::deque, which makes the transition
harder. Plus, I still think if the "middle" isn't very far from
either end, perf might be OK. You should benchmark with
a real use case.

Reviewed By: mjoras

Differential Revision: D28530405

fbshipit-source-id: 8148549c222b7787045c9e0d5dfc70a10cdcdb6e