Avoid rtt=0 due to different timestamp sources with difference < 1 us
Summary:
The current rttSample calculation allows for a bad rtt=0 sample due to loss of precision due to casting. When the ACK timestamp comes from the socket timestamp and is less than 1 us different from the packet send time, the duration_cast to microseconds results in a rtt = 0 us.
By changing this to the ceiling of the difference, the rtt=0 samples are avoided.

Reviewed By: mjoras

Differential Revision: D30347325

fbshipit-source-id: 03c585d8a879390f1f5a16157b94ccbe4e732c54