QUIC connector class
Summary:
This a QUIC connector class that is loosely based on HQConnector (https://fburl.com/code/zx8pqhf0).

This class will be used in proxygen healthchecks for connect only checks. We currently using HQConnector class but really there is no need to involve HTTP when a connect only check is needed.

This class will establish a QUIC connection up until `onTransportReady()`, report success to the HC and then will drop the connection.

Usage is in the diff above.

Reviewed By: mjoras

Differential Revision: D30490153

fbshipit-source-id: 9182a6e08643c832ac7b2972959440c926550b15